requirements.txt
astroid==2.5.6
attrs==20.3.0
autopep8==1.5.7
blessed==1.17.6
boto3==1.17.89
botocore==1.20.89
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
flake8==3.9.2
idna==2.10
importlib-metadata==4.5.0
inquirer==2.7.0
isort==5.8.0
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.8.3
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.4.1
readchar==2.0.1
requests==2.25.1
rope==0.19.0
s3transfer==0.4.2
six==1.16.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.5
venmo-api==0.3.1
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.1
